requirements.txt
# Core dependencies
PyYAML>=6.0        
numpy>=1.22.0      
pandas>=1.3.5      
matplotlib>=3.5.0  
seaborn>=0.11.2    
torch>=1.13.0      

# YOLOv8 and related
ultralytics>=8.0.0

opencv-python>=4.5.5  

# Utilities
tqdm>=4.62.3       
pathlib>=1.0.1     
